requirements.txt
# Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Authentication & Authorization
django-oauth-toolkit==2.3.0
social-auth-app-django==5.4.0
django-allauth==0.58.2
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.1
django-rest-framework-social-oauth2==1.1.0
social-auth-core==4.5.1

# AI & ML
google-generativeai==0.3.2
openai==1.8.0
numpy==1.26.3
tensorflow==2.16.1  # Latest version compatible with Python 3.12
librosa==0.10.1  # For audio processing
soundfile==0.12.1  # Required for audio processing

# Image Processing
Pillow==10.2.0
python-magic==0.4.27
python-magic-bin==0.4.14  # For Windows support

# Storage & Media
django-storages==1.14.2
boto3==1.34.14
cloudinary==1.36.0
django-cleanup==8.0.0

# Database & Caching
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.1

# Filtering & API Documentation
django-filter==23.5
drf-yasg==1.21.7
markdown==3.5.1
django-rest-swagger==2.2.0

# Location Services
geopy==2.4.1
django-location-field==2.7.2

# Background Tasks & Monitoring
celery==5.3.6
flower==2.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
supervisor==4.2.5

# Development & Deployment
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
django-debug-toolbar==4.2.0
python-decouple==3.8
dj-database-url==2.1.0

# Payment Processing
stripe==7.11.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.0
factory-boy==3.3.0
faker==22.0.0

# Security
django-ratelimit==4.1.0
django-axes==6.3.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
urllib3==2.1.0
certifi==2023.11.17
chardet==5.2.0
idna==3.6
django-phonenumber-field[phonenumbers]==7.3.0  # For phone number handling
Werkzeug==3.0.1  # Required for debugger 